Some notes I'm developing in typst.
I am using a local typst template for this.  I suppose I'll migrate the
material to Rmd at some point, and combine it with the .Rmd file that is
in this directory. For now, though, I'm liking how `typst watch` works;
in combination with using `skim`, this yields essentially instantaneous
updates to the PDF. I'm surprised at how pleasant that is.